bugfix
Buffer decryption routine used to look at the last byte of the plaintext
to determine the padding size.
If used with a wrong key, the last byte could be, say, 0xff, which is
definitely not the right padding size.
Now it checks for proper padding at the end of the plaintext and drops
the last block if no valid padding was found.